can now generate shitty procedural terrain

+++ include/Game/Scenes/GameScene.hpp
@@ -12,6 +12,6 @@ public:
 private:
     Entity *player = nullptr;
     std::vector<Entity *> _blocks;
-    std::vector<std::vector<float>> _heightMap;
+    std::vector<std::vector<int>> _heightMap;
     bool load = false;
 };
